requirements.txt
deepface==0.0.79
tensorflow==2.15.1  # Newer LTS version compatible with deepface
keras==2.15.0       # Matches TF 2.15.x
opencv-python==4.8.0.76
matplotlib==3.7.1

weaviate-client==3.26.7   # ✅ fully compatible with Weaviate 1.25.3
gradio==3.50.2  # More stable version
pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0